Move OrderEvent model to a events package and prepare the structure
