+++ saleor/events/models/__init__.py
@@ -1,0 +1,4 @@
+from .OrderEvent import OrderEvent
+
+__all__ = [
+    "OrderEvent"]
