+++ saleor/graphql/order/mutations/draft_orders.py
@@ -5,7 +5,8 @@ from graphene.types import InputObjectType
 from ....account.models import User
 from ....core.exceptions import InsufficientStock
 from ....core.utils.taxes import ZERO_TAXED_MONEY
-from ....order import OrderEvents, OrderStatus, models
+from ....events.types import OrderEvents
+from ....order import OrderStatus, models
 from ....order.utils import (
     add_variant_to_order, allocate_stock, change_order_line_quantity,
     delete_order_line, recalculate_order)
