+++ saleor/graphql/order/types.py
@@ -3,6 +3,7 @@ import graphene_django_optimizer as gql_optimizer
 from django.core.exceptions import ValidationError
 from graphene import relay
 
+from ...events import models as events_models
 from ...order import models
 from ...order.models import FulfillmentStatus
 from ...product.templatetags.product_images import get_product_image_thumbnail
@@ -40,7 +41,7 @@ class OrderEvent(CountableDjangoObjectType):
 
     class Meta:
         description = 'History log of the order.'
-        model = models.OrderEvent
+        model = events_models.OrderEvent
         interfaces = [relay.Node]
         only_fields = ['id']
