+++ saleor/settings.py
@@ -221,6 +221,7 @@ INSTALLED_APPS = [
     'saleor.core',
     'saleor.graphql',
     'saleor.menu',
+    'saleor.events',
     'saleor.order',
     'saleor.dashboard',
     'saleor.seo',
